elementalconductor: use a pointer for Job.ContentDuration
When creating a new job, we don't want to send this data, only when
retrieving the job. omitempty will only be respected if we use a
pointer. There's probably a better solution, like having different
structs for CreateJob and GetJob, but I'm not sure if this is really a
better solution.